"fix" for a weird inter-dep between configs
